backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
python-dotenv==1.1.0
pydantic==2.11.4
pydantic-settings==2.9.1
langchain==0.3.25
langchain-google-genai==2.1.4
langchain-experimental==0.3.4
pandas==2.2.3
matplotlib==3.10.3
